# openapi/SmartGrao.WebApi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SmartGrao.WebApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/growth-stages": {
      "get": {
        "tags": [
          "GrowthStages"
        ],
        "operationId": "GetGrowthStages",
        "parameters": [
          {
            "name": "crop",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Crop"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GrowthStageOptionViewModel"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/protocols": {
      "get": {
        "tags": [
          "Protocols"
        ],
        "operationId": "GetProtocols",
        "parameters": [
          {
            "name": "crop",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Crop"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProtocolStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProtocolSummaryViewModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Protocols"
        ],
        "operationId": "CreateProtocol",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProtocolViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProtocolViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProtocolViewModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/protocols/{id}": {
      "get": {
        "tags": [
          "Protocols"
        ],
        "operationId": "GetProtocolById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/protocols/{id}/versions": {
      "post": {
        "tags": [
          "Protocols"
        ],
        "operationId": "CreateProtocolVersion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/protocols/{id}/name": {
      "put": {
        "tags": [
          "Protocols"
        ],
        "operationId": "RenameProtocol",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameProtocolViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameProtocolViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RenameProtocolViewModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/protocols/{id}/items": {
      "post": {
        "tags": [
          "Protocols"
        ],
        "operationId": "AddProtocolItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProtocolItemViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProtocolItemViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddProtocolItemViewModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/protocols/{id}/items/{itemId}": {
      "delete": {
        "tags": [
          "Protocols"
        ],
        "operationId": "RemoveProtocolItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/protocols/{id}/publication": {
      "put": {
        "tags": [
          "Protocols"
        ],
        "operationId": "PublishProtocol",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/protocols/{id}/retirement": {
      "put": {
        "tags": [
          "Protocols"
        ],
        "operationId": "RetireProtocol",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolViewModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddProtocolItemViewModel": {
        "required": [
          "targetId",
          "organ",
          "unit",
          "photosRequested",
          "instructions",
          "referenceLevel"
        ],
        "type": "object",
        "properties": {
          "targetId": {
            "type": "string",
            "format": "uuid"
          },
          "organ": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PlantOrgan"
              }
            ]
          },
          "unit": {
            "$ref": "#/components/schemas/CountUnit"
          },
          "photosRequested": {
            "type": "integer",
            "format": "int32"
          },
          "instructions": {
            "type": "string"
          },
          "referenceLevel": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReferenceLevelViewModel"
              }
            ]
          }
        }
      },
      "AutomationCapability": {
        "enum": [
          "ManualRecord",
          "UnderValidation",
          "AutomationEnabled"
        ]
      },
...
      "CountUnit": {
        "enum": [
          "Undefined",
          "AttackedPlantPercentage",
          "AttackedPlantsPer10MetreRow",
          "InsectsPerPlant",
          "InsectsPerTenPlants",
          "InsectsPerTrap",
          "LesionedLeafAreaPercentage",
          "SeverityScore1To9",
          "Presence"
        ]
      },
...
      "CreateProtocolViewModel": {
        "required": [
          "code",
          "crop",
          "name"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "crop": {
            "$ref": "#/components/schemas/Crop"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "Crop": {
        "enum": [
          "Undefined",
          "Soybean",
          "Corn",
          "Cotton",
          "Coffee",
          "Sugarcane",
          "Wheat",
          "Beans",
          "Rice",
          "Sorghum",
          "Other"
        ]
      },
...
      "GrowthStageOptionViewModel": {
        "required": [
          "code",
          "description"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "PlantOrgan": {
        "enum": [
          "Undefined",
          "Whorl",
          "Leaf",
          "EarLeaf",
          "StalkBase",
          "Stalk",
          "Root",
          "Ear",
          "Tassel",
          "Silk",
          "Seedling",
          "WholePlant",
          null
        ]
      },
...
      "ProtocolItemViewModel": {
        "required": [
          "id",
          "targetId",
          "targetCode",
          "targetCommonName",
          "targetAutomation",
          "organ",
          "unit",
          "photosRequested",
          "instructions",
          "referenceLevel"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "targetId": {
            "type": "string",
            "format": "uuid"
          },
          "targetCode": {
            "type": "string"
          },
          "targetCommonName": {
            "type": "string"
          },
          "targetAutomation": {
            "$ref": "#/components/schemas/AutomationCapability"
          },
          "organ": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PlantOrgan"
              }
            ]
          },
          "unit": {
            "$ref": "#/components/schemas/CountUnit"
          },
          "photosRequested": {
            "type": "integer",
            "format": "int32"
          },
          "instructions": {
            "type": "string"
          },
          "referenceLevel": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReferenceLevelViewModel"
              }
            ]
          }
        }
      },
      "ProtocolStatus": {
        "enum": [
          "Draft",
          "Published",
          "Retired"
        ]
      },
      "ProtocolSummaryViewModel": {
        "required": [
          "id",
          "code",
          "version",
          "crop",
          "name",
          "status",
          "targetCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "crop": {
            "$ref": "#/components/schemas/Crop"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ProtocolStatus"
          },
          "targetCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ProtocolViewModel": {
        "required": [
          "id",
          "code",
          "version",
          "crop",
          "name",
          "status",
          "items"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "crop": {
            "$ref": "#/components/schemas/Crop"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ProtocolStatus"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProtocolItemViewModel"
            }
          }
        }
      },
...
      "ReferenceLevelViewModel": {
        "required": [
          "threshold",
          "source"
        ],
        "type": "object",
        "properties": {
          "threshold": {
            "type": "number",
            "format": "double"
          },
          "source": {
            "type": "string"
          }
        }
      },
      "RenameProtocolViewModel": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}